+++ src/TwigComponent/src/Event/MountEvent.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\UX\TwigComponent\Event;
+
+use Symfony\Contracts\EventDispatcher\Event;
+use Symfony\UX\TwigComponent\ComponentMetadata;
+
+/**
+ * @author Simon André <smn.andre@gmail.com>
+ */
+class MountEvent extends ComponentEvent
+{
+}